refact: feed hook, ui분리
