+++ src/pages/home/HomeFeed.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect, useRef } from 'react';
-import Feed from './feed/Feed'; 
 import styles from "./HomeFeed.module.css";
 import axios from 'axios';
 import SlideBtn from '../../Components/SlideButton'; 
 import { useNavigate } from 'react-router';
+import Feed from '../../Components/feed/Feed'; // Feed 컴포넌트 import
 
 const getFeeds = async (navigate) => {
     const token = localStorage.getItem("accessToken");
@@ -33,7 +33,6 @@ const getFeeds = async (navigate) => {
 
 function FeedList() {
     const navigate = useNavigate();
-
     const [feeds, setFeeds] = useState([]);
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState(null);
@@ -52,7 +51,7 @@ function FeedList() {
         };
 
         fetchPosts();
-    }, []);
+    }, [navigate]);
 
     if (loading) {
         return <div>로딩 중...</div>;
@@ -66,26 +65,26 @@ function FeedList() {
         <div className={styles.feedScroll}>
             <div className={styles.homeFeedTitle}>
                 <div className={styles.homeTitle}>팔로잉 스타일 모아보기</div>
-                <div className={styles.homeSubTitle}>팔로우한 사람들의 최신 스타일을 확인하세요.<br/>
-                    마음에 드는 스타일을 클릭해 아이템 정보를 확인해보세요!</div>
+                <div className={styles.homeSubTitle}>
+                    팔로우한 사람들의 최신 스타일을 확인하세요.<br />
+                    마음에 드는 스타일을 클릭해 아이템 정보를 확인해보세요!
+                </div>
             </div>
             <div className={styles.feedList} ref={feedListRef}>
-                {feeds
-                    .map(feed => {
-                        const mediaUrls = Array.isArray(feed.media_url_list) ? feed.media_url_list : [];
-                        return (
-                            <Feed
-                                key={feed.media_id}
-                                media_url_list={mediaUrls}
-                                profile_url={feed.profile_url}
-                                username={feed.username}
-                                media_id={feed.media_id}
-                                currentIndex={0}
-                                home={true}
-                            />
-                        );
-                    })
-                }
+                {feeds.map(feed => {
+                    const mediaUrls = Array.isArray(feed.media_url_list) ? feed.media_url_list : [];
+                    return (
+                        <Feed
+                            key={feed.media_id}
+                            media_url_list={mediaUrls}
+                            profile_url={feed.profile_url}
+                            username={feed.username}
+                            media_id={feed.media_id}
+                            currentIndex={0}
+                            home={true}
+                        />
+                    );
+                })}
                 <div style={{ height: '10px' }} />
             </div>
             <SlideBtn /> {/* SlideButton 렌더링 */}
